refactor CPID so we don;t make a web request in a constructor
